scanmed application refactor
moved logic from ScanmedController to ScanmedService
switched from field autowiring to AllArgsConstructor autowiring
