+++ scanmed/src/test/java/software/xsolve/springcloud/scanmed/ApplicationTest.java
@@ -1,0 +1,16 @@
+package software.xsolve.springcloud.scanmed;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class ApplicationTest {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
